requirements.txt
streamlit>=1.31.0
google-genai>=1.0.0
duckduckgo-search>=4.1.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
